Chapter14/14_2_benchmark_with_bgperf/pip-requirements.txt
docker
pyyaml
pyroute2
nsenter
netaddr
mako
six
packaging
